ECOCSS-3736, ECOCSS-3737, ECOCSS-3738, ECOCSS-3739, ECOCSS-3740, ECOCSS-3741, ECOCSS-3742, ECOCSS-3743, ECOCSS-3744, ECOCSS-3745, ECOCSS-3746, ECOCSS-3747, ECOCSS-3748, ECOCSS-3749, ECOCSS-3750, ECOCSS-3751, ECOCSS-3752, ECOCSS-3753, ECOCSS-3754, ECOCSS-3755, ECOCSS-3756, ECOCSS-3757, ECOCSS-3758, ECOCSS-3759, ECOCSS-3760, ECOCSS-3761, ECOCSS-3762, ECOCSS-3763, ECOCSS-3764, ECOCSS-3765, ECOCSS-3766, ECOCSS-3767, ECOCSS-3768, ECOCSS-3769, ECOCSS-3770, ECOCSS-3771, ECOCSS-3772, ECOCSS-3773, ECOCSS-3774, ECOCSS-3775, ECOCSS-3776, ECOCSS-3777, ECOCSS-3778, ECOCSS-3779, ECOCSS-3780, ECOCSS-3781, ECOCSS-3782, ECOCSS-3783, ECOCSS-3784, ECOCSS-3785, ECOCSS-3786, ECOCSS-3787, ECOCSS-3788, ECOCSS-3789, ECOCSS-3790, ECOCSS-3791, ECOCSS-3792, ECOCSS-3793, ECOCSS-3794, ECOCSS-3795, ECOCSS-3796, ECOCSS-3797, ECOCSS-3798, ECOCSS-3799, ECOCSS-3800, ECOCSS-3801, ECOCSS-3802, ECOCSS-3803, ECOCSS-3804, ECOCSS-3805: fix API bugs

+++ ReservationService/src/main/java/com/example/reservationservice/dto/FairnessMemberDTO.java
@@ -3,7 +3,7 @@ package com.example.reservationservice.dto;
 import java.time.LocalDateTime;
 
 public class FairnessMemberDTO {
-    private Integer userId;
+    private Long userId;
     private String fullName;
     private String email;
     private Double ownershipPercentage;
@@ -17,7 +17,7 @@ public class FairnessMemberDTO {
 
     public FairnessMemberDTO() {}
 
-    public FairnessMemberDTO(Integer userId, String fullName, String email, Double ownershipPercentage, 
+    public FairnessMemberDTO(Long userId, String fullName, String email, Double ownershipPercentage, 
                              Double usageHours, Double usagePercentage, Double difference, 
                              Double fairnessScore, String priority, LocalDateTime lastUsageEnd, 
                              LocalDateTime nextReservationStart) {
@@ -39,8 +39,8 @@ public class FairnessMemberDTO {
         return new FairnessMemberDTOBuilder();
     }
 
-    public Integer getUserId() { return userId; }
-    public void setUserId(Integer id) { this.userId = id; }
+    public Long getUserId() { return userId; }
+    public void setUserId(Long id) { this.userId = id; }
     public String getFullName() { return fullName; }
     public void setFullName(String n) { this.fullName = n; }
     public String getEmail() { return email; }
@@ -64,7 +64,7 @@ public class FairnessMemberDTO {
 
     public static class FairnessMemberDTOBuilder {
         private FairnessMemberDTO instance = new FairnessMemberDTO();
-        public FairnessMemberDTOBuilder userId(Integer v) { instance.setUserId(v); return this; }
+        public FairnessMemberDTOBuilder userId(Long v) { instance.setUserId(v); return this; }
         public FairnessMemberDTOBuilder fullName(String v) { instance.setFullName(v); return this; }
         public FairnessMemberDTOBuilder email(String v) { instance.setEmail(v); return this; }
         public FairnessMemberDTOBuilder ownershipPercentage(Double v) { instance.setOwnershipPercentage(v); return this; }
